Blocking out the method used to handle world based commands (spells). A few issues are popping up e.g. can't test distances for spells without the actual cry components. Code only executes a crylog operation so far. Don't like how repetitive the handling of basically the same thing is - tempted to just do them all as a single struct with an enum to differentiate them. Don't like the performance implication of having to check for casts of every spell every frame.
Tossed up the idea of adding the spell cast as a func using the meta for the component, but not sure if I can iterate all components for an entity and if I can access the stored func without know the exact component. Went round and round in my head and finally decided I would just blunt force it for now. If I end up with 100 components this will definitely be a bad way to handle it.